+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/AlbumTablePanel.java
@@ -83,7 +83,7 @@ public final class AlbumTablePanel extends EntityTablePanel {
 			dialog.toFront();
 		}
 		else {
-			Dialogs.componentDialog()
+			Dialogs.dialog()
 							.component(coverPanel)
 							.owner(Utilities.parentWindow(this))
 							.title(title)