+++ demos/employees/src/main/java/is/codion/demos/employees/ui/DepartmentTablePanel.java
@@ -68,7 +68,7 @@ public class DepartmentTablePanel extends EntityTablePanel {
 		JasperPrint employeeReport = tableModel().connection()
 						.report(Employee.EMPLOYEE_REPORT, reportParameters);
 
-		Dialogs.componentDialog()
+		Dialogs.dialog()
 						.component(new JRViewer(employeeReport))
 						.owner(this)
 						.modal(false)