+++ demos/manual/src/main/java/is/codion/manual/common/demo/ApplicationPanel.java
@@ -428,7 +428,7 @@ public final class ApplicationPanel extends JPanel {
 
 		ApplicationPanel applicationPanel = new ApplicationPanel(applicationModel);
 
-		Dialogs.componentDialog()
+		Dialogs.dialog()
 						.component(applicationPanel)
 						.title("Codion Input Components Demo")
 						.icon(Logos.logoTransparent())