+++ demos/manual/src/main/java/is/codion/manual/quickstart/Example.java
@@ -112,7 +112,7 @@ public final class Example {
 		//populate the table model with data from the database
 		customerModel.tableModel().items().refresh();
 
-		Dialogs.componentDialog()
+		Dialogs.dialog()
 						.component(customerPanel)
 						.title("Customers")
 						.show();