+++ demos/world/src/main/java/is/codion/demos/world/ui/CountryTablePanel.java
@@ -67,7 +67,7 @@ final class CountryTablePanel extends EntityTablePanel {
 	}
 
 	private void viewReport(JasperPrint countryReport) {
-		Dialogs.componentDialog()
+		Dialogs.dialog()
 						.component(new JRViewer(countryReport))
 						.owner(this)
 						.modal(false)