+++ swing/framework-ui/src/main/java/is/codion/swing/framework/ui/component/EntitySearchField.java
@@ -604,7 +604,7 @@ public final class EntitySearchField extends HintTextField {
 
 	private JPopupMenu createPopupMenu() {
 		return menu(Controls.controls(Control.builder()
-						.command(() -> Dialogs.componentDialog()
+						.command(() -> Dialogs.dialog()
 										.component(settingsPanel())
 										.owner(EntitySearchField.this)
 										.title(FrameworkMessages.settings())